# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Fratelanza Management Console API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /templates:
    get:
      operationId: listTemplates
      tags: [templates]
      summary: List all service templates
      parameters:
        - in: query
          name: category
          schema:
            type: string
      responses:
        "200":
          description: Templates list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Template"
    post:
      operationId: createTemplate
      tags: [templates]
      summary: Add a new service template
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TemplateInput"
      responses:
        "201":
          description: Created template
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Template"
  /templates/{id}:
    patch:
      operationId: updateTemplate
      tags: [templates]
      summary: Update a service template
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TemplateUpdate"
      responses:
        "200":
          description: Updated template
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Template"
        "404":
          description: Not found
    delete:
      operationId: deleteTemplate
      tags: [templates]
      summary: Delete a service template
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found

  # ── Sales Quotes ──────────────────────────────────────────────────────
  /quotes:
    get:
      operationId: listQuotes
      tags: [quotes]
      summary: List sales quotes
      parameters:
        - in: query
          name: client
          schema:
            type: string
        - in: query
          name: search
          schema:
            type: string
      responses:
        "200":
          description: Quotes list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Quote"
    post:
      operationId: createQuote
      tags: [quotes]
      summary: Create a new sales quote
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/QuoteInput"
      responses:
        "201":
          description: Created quote
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Quote"
  /quotes/{id}:
    patch:
      operationId: updateQuote
      tags: [quotes]
      summary: Update a sales quote
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/QuoteUpdate"
      responses:
        "200":
          description: Updated quote
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Quote"
        "404":
          description: Not found
    delete:
      operationId: deleteQuote
      tags: [quotes]
      summary: Delete a quote
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found

  # ── General Expenses ──────────────────────────────────────────────────
...
  /tasks:
    get:
      operationId: listTasks
      tags: [tasks]
      summary: List tasks with optional filters
      parameters:
        - in: query
          name: status
          schema:
            type: string
        - in: query
          name: projectId
          schema:
            type: integer
      responses:
        "200":
          description: Tasks list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Task"
    post:
      operationId: createTask
      tags: [tasks]
      summary: Create a new task
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TaskInput"
      responses:
        "201":
          description: Created task
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
  /tasks/{id}:
    patch:
      operationId: updateTask
      tags: [tasks]
      summary: Update a task (status, details, etc.)
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TaskUpdate"
      responses:
        "200":
          description: Updated task
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
        "404":
          description: Not found
    delete:
      operationId: deleteTask
      tags: [tasks]
      summary: Delete a task
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found

  # ── Finance / P&L Report ──────────────────────────────────────────────
...
components:
  schemas:
...
    Template:
      type: object
      required: [id, category, name, cost]
      properties:
        id:
          type: integer
        category:
          type: string
        name:
          type: string
        cost:
          type: number
        expenses:
          type: number
        multiplier:
          type: number
        broker:
          type: number
        students:
          type: integer
    TemplateInput:
      type: object
      required: [category, name, cost]
      properties:
        category:
          type: string
        name:
          type: string
        cost:
          type: number
        expenses:
          type: number
        multiplier:
          type: number
        broker:
          type: number
        students:
          type: integer
    TemplateUpdate:
      type: object
      properties:
        category:
          type: string
        name:
          type: string
        cost:
          type: number
        expenses:
          type: number
        multiplier:
          type: number
        broker:
          type: number
        students:
          type: integer

    # ── Quote ──────────────────────────────────────────────────────────
    Quote:
      type: object
      required: [id, clientName, price, date]
      properties:
        id:
          type: integer
        clientName:
          type: string
        projectName:
          type: ["string", "null"]
        price:
          type: number
        language:
          type: ["string", "null"]
        date:
          type: string
        paymentTerms:
          type: ["string", "null"]
        milestones:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
    QuoteInput:
      type: object
      required: [clientName, price]
      properties:
        clientName:
          type: string
        projectName:
          type: string
        price:
          type: number
        language:
          type: string
        date:
          type: string
        paymentTerms:
          type: string
        milestones:
          type: string
        notes:
          type: string
    QuoteUpdate:
      type: object
      properties:
        clientName:
          type: string
        projectName:
          type: string
        price:
          type: number
        language:
          type: string
        paymentTerms:
          type: string
        milestones:
          type: string
        notes:
          type: string

    # ── Expense ────────────────────────────────────────────────────────
...
    Task:
      type: object
      required: [id, title, status, createdAt]
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: ["string", "null"]
        status:
          type: string
        priority:
          type: ["string", "null"]
        projectId:
          type: ["integer", "null"]
        projectName:
          type: ["string", "null"]
        assignedTo:
          type: ["string", "null"]
        dueDate:
          type: ["string", "null"]
        createdAt:
          type: string
    TaskInput:
      type: object
      required: [title]
      properties:
        title:
          type: string
        description:
          type: string
        status:
          type: string
        priority:
          type: string
        projectId:
          type: integer
        assignedTo:
          type: string
        dueDate:
          type: string
    TaskUpdate:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        status:
          type: string
        priority:
          type: string
        projectId:
          type: integer
        assignedTo:
          type: string
        dueDate:
          type: string

    # ── Finance Report ─────────────────────────────────────────────────
...