# lib/api-zod/src/generated/types/templateInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Fratelanza Management Console API
 * OpenAPI spec version: 0.1.0
 */

export interface TemplateInput {
  category: string;
  name: string;
  cost: number;
  expenses?: number;
  multiplier?: number;
  broker?: number;
  students?: number;
}